Fix dynamic entity auto-loading on initial map.
The previous system of 'stale entities' running frequently to detect
if a call had been made to change the map, and if it had, to call
get_entities in the entity manager (which would then ask the map
manager) was tailored to an earlier version of the code that wasn't
so capable of bi-directional talk between managers.

What was happening here was that the first map would be set to loading,
it would load, change maps would be called, the init() would run, and
then stale_entities() would somehow get hit after and wipe everything
that wasn't part of the map's definition out. Rather than spend time
tracing stale_entities out and coaxing it into behaving better, the
above solution of just excising it entirely seemed to fit.

In order to make this change, some related changes had to be made to
is_loading, change_maps, and get_entities, to block bad reads when the
map hasn't loaded yet more smartly and to directly make the entity load
call from the map manager across to the entity manager. It is a bit
more coupled now, but I think that's manageable.