Add initial wave collapse work.
INTRO

So, this is like a markov bot for map generation.

The idea is taken from https://github.com/mxgmn/WaveFunctionCollapse,
and just reimplemented partly for the fun of it, and partly because
neither mxgmn's nor any of the ports of their algorithm looked like
they worked quite the way I wanted (they all use reflections and
rotations of tiles, etc.), and the code in them is very tough to
follow -- a lot of responsibilities are mixed, and I think that I can
(with time) make more readable code that accomplishes something
similar.

THE ALGORITHM

First, it reads an entire map of tiles to count the frequencies of
each individual tile type and to count the frequencies of each tile
occurring in each direction of each tile. For example, the frequency
of grass tiles occurring above water tiles.

It places these in two objects, 'frequencies' and 'dir_frequencies',
keyed as 'tile_type -> frequency' and 'tile_type.direction.tile_type ->
frequency', e.g. 'grass -> 0.5' or 'grass.above.water -> 0.5', where
the first number indicates 50% of the map is grass, and thus that a
given tile has a 50% chance of being grass, and the second number
indicates that 50% of the tiles above grass tiles are water tiles,
and thus that a given tile above a grass tile has a 50% chance of
being water during generation.

A field of valueless nodes is then created of a given output size,
and the frequency and dir_frequencies are passed into them. Each node
carries its own internal copy of these lists, and modifies them as work
is done. The nodes are placed into what is here called a "selection
list", which is just a bucket-array where the bucket values are "how
constrained the nodes in this bucket are". Initially, no nodes have
any constraints, so they are all in one bucket.

The actual wavecollapse algorithm works by choosing a random node
from the bucket representing the most-constrained nodes, selecting
a value for it randomly from its internally-held frequency map, and
then visiting each neighbouring node (that has not already been given
a value), applying any new constraints that result from its new
neighbour, and altering its internally held frequencies according to
that neighbour's directional frequency list for this particular case.

For example, if you have a blank node to the right of a water node, and
water is never beside mountains, you will write a new constraint for
the blank node that it cannot be a mountain. Then, proceeding to
directions, if nodes to the right of water can be water 90% of the
time, grass 10% of the time, and mountains 0% of the time (this is
imagining a map with just these three tile types), you set the blank
node's internal frequencies to those values.

If the blank node already had a neighbour, it will have already had its
frequencies set by that neighbour, and we don't wish to just forget
about it! So we keep track of the number of contributors to the node's
frequencies, and in the event it already has some, we average the new
values with the old on a footing that gives all contributors equal
weight, by multiplying the existing frequencies by the number of nodes
already contributing, then adding, then dividing by the new number of
contributors.

In the act of constraining the nodes, the algorithm plucked each blank
node from its bucket in the selection list, then placed it into its new
bucket after applying them. By this method, nodes migrate to positions
of higher constrainedness, and the algorithm then operates on the most
constrained nodes first without spending much time on sorting.

One unusual aspect of the constraint/frequency bookkeeping in this
version of the algorithm is that, in the event of a contradiction is
imminent -- that is, that the last possibility for a node is about to
be constrained, we just decline to do so. The more common alternative
is to restart map generation at the point a contradiction is hit, but
I found that random tileset seeds often generate distributions that
lead to contradictions, and feel the cost of just taking the least-
objectionable-tile is low enough (it makes some neat maps!) that it's
the right move here.

ODDITIES / REMAINING TODOS

There are superfluous comments and debug statements sprinkled through
this whole thing right now. There's also a lot of potential work to be
done in simplifying the handling of the constraints -- the 'cruft' var
in the pick_by_distribution function has a very uncertain use at this
point in time -- it's sort of become a blanket fudge-factor for any
errors in probability-calculation, which isn't its intended purpose.

Breaking the frequency generation and the map generation from the
frequencies apart a bit more would be great, because it would allow
hand-rolling of frequencies and re-use of past frequency sets that
yielded good results. It also would be way faster to generate each
new map without first building a whole random map and reading it to
generate a distribution!

There are also probably some significant data-type and algorithm wins
waiting in the wings to reduce both memory-use and runtime in the
generation phase. There are a lot of superfluous keys, copies of whole
objects, and overuse of dictionaries here, which I am excited to try
to fix up.