Generate maps with wavecollapse.
This is a mess, in that it's the first moment that things are
working properly and I want to save state. More to come that will
fix things up.

Generates a large seed map with "features" of blocks of water,
grass, brush, and forest, and a long road, then uses that to build
the wavecollapse distribution and generate the wavecollapsed map.

Takes forever to build. Runs twice at startup for some reason(?),
but makes some _cool freaking maps_.