For posterity: wavecollapse pre-directional freqs.
This version of the wavecollapse code only used the frequency of
node placement in the whole map, not the frequency of nodes next to
other nodes, and it produced very noticeably subpar results. I thought
I'd lost this version, but now that it's found, we can keep it in the
commit log for no one to ever care about.