Simplify creek draw loop down to nothing.
It will be entirely dependent upon computer speed for now, but all
attempts to make it intelligently run a few updates per draw or a few
draws per update when needed to achieve a fixed framerate have had a
ruinous impact on the framerate in general and don't seem to work the
way I expect.

So, we'll defer that -- the plain update/draw loop is working fine for
now.