Add an index-by-item-id to quadtree.
Speeds up removal by id and get by id to O(1), and speeds entity
movement up to O(logN), from O(N) on all of these operations, at
the cost of storing an additional pointer to each entity in a
dictionary, and storing a link to the local node's storage pool
in each entity.

Previously, a full tree scan had been required to find an element
by id, which is how moves, get by id, and removal-by-id were each
performed. Now, with the entity index, an entity can be got straight
out of the dictionary from its id and it can be swiftly removed via
its own pointer to its local tree-node's storage pool, eliminating
the need for a logN scan of the tree by coordinate to find the pool
it must be eliminated from.

Movement still requires reinsertion after removal, which remains logN
in walking the tree by coordinate-search.

It's a bit messy to put the pointers to local-node-storage-pools into
the objects being stored in the tree -- very messy, actually! But I'm
the only one using this, so it's fine. The fix to that which I may
write in the future is to store a second index in the tree, or to add
a layer of abstraction to the existing index, where these pointers can
be kept and managed.

This change was prompted by reading performance flame charts and seeing
massive slowdowns as soon as I had >5-10 moving objects on the screen
alongside N hundred to 1-2 thousand entities stored in the tree -- it
was way too many O(N) scans on way too big an N. Frame rates (once I
had fixed some frame-rate-juggling behaviour in the inner game loop)
jumped from 4-5fps to often 20-30fps, much to my joy. :D