Add dashing, improve movement, show help at start.
Dashing! When you can only move one square at a time and the map is
2 screen widths by 2 screen heights in size, it's a bit of a slog to
get around. Dashes are fixed-length sprints the hero can do which
require you to stop at their end before you can do a normal move or
a new dash.

Dashing into monsters or treasure will deal extra damage, but will
terminate the dash, requiring the hero to initiate another dash or
to resume regular movement. Dashing into other objects will destroy
them and stop the hero -- it destroys a tree outright instead of
turning it into a stump/dead tree first, at the cost of sacrificing
the wood gain you'd normally get from removing a tree.

I think it would be useful to add a cooldown or stamina mechanism to
avoid constantly dashing everywhere, with powerups to increase your
stamina or reduce the cooldown, in addition to powerups to increase
the dash attack power, stop objects from terminating dash, or add to
the dash length. More powerful enemeies are likely required to make
those things feel useful.

Because this is instrumental to having fun and it's not obvious, I
chose to show the help modal by default and add instructions to the
modal on its dismissal / update its information. It also required a
bit of general fixing, as it seemed to want to sometimes resist being
removed -- an indication of a movement-related bug I've seen a few
times now and will have to fix soon.