Optionally draw from local canvases of given size.
If you know the size you want an image to be in a game, and where it
resides in a larger tileset, there's no  reason to grab the wrong-sized
version of it out of the source image every time you render a frame --
that's quite expensive!

Of course, if you want to you should be able to do that, but this
change allows the config to specify a width and height of a resource,
and by default draws resources out of their source image onto their
own canvases of their specified width and height, so that the render
code can just blit the whole canvas' contents to a given position,
without any extra work.

If the width or height weren't specified, or if the tile has a custom
scaling factor that would alter the scale the resource is drawn at, the
more complex (and as I've found, slower) rendering function is used.