Add Time and Data managers to Creek.
TimeManager keeps track of time globally, advancing each frame, and
provides facilities for starting/stopping/querying named timers.

DataManager gives an interface to a global data object for when you
want to save some state that persists between frames and don't want
to hang it off of a particular map or the player or a different odd
object that's floating around.

This change also makes some light use of the Time and Data managers
in the run loop which will likely change in the future.